+++ pkg/container/podman_container_test.go
@@ -13,13 +13,13 @@ func TestPodmanContainerInspect(t *testing.T) {
 	rawOutput := []byte(`{"Test Key":"Test Value"}`)
 
 	// Mock dependencies
-	oldFunc := executePodmanInspectFunc
-	executePodmanInspectFunc = func(string) (stdout, stderr []byte, err error) {
+	oldFunc := podmanFunc
+	podmanFunc = func(...string) (stdout, stderr []byte, err error) {
 		stdout = rawOutput
 		return stdout, stderr, err
 	}
 	defer func() {
-		executePodmanInspectFunc = oldFunc
+		podmanFunc = oldFunc
 	}()
 
 	podmanContainer := &PodmanContainer{
@@ -35,13 +35,13 @@ func TestPodmanContainerInspect(t *testing.T) {
 func TestPodmanContainerInspectError(t *testing.T) {
 	testError := errors.New("fake error")
 	// Mock dependencies
-	oldFunc := executePodmanInspectFunc
-	executePodmanInspectFunc = func(string) (stdout, stderr []byte, err error) {
+	oldFunc := podmanFunc
+	podmanFunc = func(...string) (stdout, stderr []byte, err error) {
 		err = testError
 		return stdout, stderr, err
 	}
 	defer func() {
-		executePodmanInspectFunc = oldFunc
+		podmanFunc = oldFunc
 	}()
 
 	podmanContainer := &PodmanContainer{
@@ -51,4 +51,52 @@ func TestPodmanContainerInspectError(t *testing.T) {
 	inspectResult, err := podmanContainer.Inspect()
 	assert.Error(t, testError, err)
 	assert.Nil(t, inspectResult)
+}
+
+func TestPodmanContainerExec(t *testing.T) {
+	standardOut := []byte("test standard output")
+	standardErr := []byte("test standard error")
+
+	// Mock dependencies
+	oldFunc := podmanFunc
+	podmanFunc = func(...string) (stdout, stderr []byte, err error) {
+		stdout = standardOut
+		stderr = standardErr
+		return stdout, stderr, err
+	}
+	defer func() {
+		podmanFunc = oldFunc
+	}()
+
+	podmanContainer := &PodmanContainer{
+		ContainerID: "test-container",
+	}
+
+	execStdout, execStderr, err := podmanContainer.Exec("test")
+	assert.NoError(t, err)
+	assert.Equal(t, standardOut, execStdout)
+	assert.Equal(t, standardErr, execStderr)
+}
+
+func TestPodmanContainerExecError(t *testing.T) {
+	testError := errors.New("fake error")
+
+	// Mock dependencies
+	oldFunc := podmanFunc
+	podmanFunc = func(...string) (stdout, stderr []byte, err error) {
+		err = testError
+		return stdout, stderr, err
+	}
+	defer func() {
+		podmanFunc = oldFunc
+	}()
+
+	podmanContainer := &PodmanContainer{
+		ContainerID: "test-container",
+	}
+
+	stdout, stderr, err := podmanContainer.Exec("test")
+	assert.Error(t, testError, err)
+	assert.Nil(t, stdout)
+	assert.Nil(t, stderr)
 }